+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.4.16 (Unreleased)
 - [Enhancement] Introduce `Consumer#wrap` for connection pooling management and other wrapped operations.
+- [Fix] Fix `kafka_format` `KeyError` that occurs when a non-hash is assigned to the kafka scope of the settings.
 
 ## 2.4.15 (2024-12-04)
 - [Fix] Assignment tracker current state fetch during a rebalance loop can cause an error on multi CG setup.